requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161728/work
certifi @ file:///Users/ec2-user/croot/certifi_1767659142737/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832772717/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744841364/work
click==8.3.1
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.1
ffmpy==1.0.0
filelock @ file:///Users/ec2-user/croot/filelock_1769440497507/work
fsspec==2026.2.0
ftfy==6.3.1
gmpy2 @ file:///opt/miniconda3/conda-bld/gmpy2_1764747893574/work
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna @ file:///opt/miniconda3/conda-bld/idna_1761912002779/work
Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2cnn4kenrm/croot/jinja2_1741710859444/work
markdown-it-py==4.0.0
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1f_uj4vxik/croot/markupsafe_1738584045311/work
mdurl==0.1.2
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
networkx @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_e6lt1dcyid/croot/networkx_1737040482044/work
numpy==1.26.4
open_clip_torch==3.3.0
orjson==3.11.7
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas==2.3.3
pillow @ file:///Users/ec2-user/croot/pillow_1772475245795/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496094255/work
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753011241/work
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML @ file:///opt/miniconda3/conda-bld/pyyaml_1763465492983/work
regex==2026.2.28
requests @ file:///opt/miniconda3/conda-bld/requests_1762359594551/work
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.13.1
timm==1.0.25
tomlkit==0.13.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
tzdata==2025.3
urllib3 @ file:///Users/ec2-user/croot/urllib3_1767810408374/work
uvicorn==0.42.0
wcwidth==0.6.0
